backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pymupdf==1.23.8
python-docx==1.1.0
rapidfuzz==3.6.1
